requirements.txt
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
APScheduler==3.10.4
dateparser==1.2.0
parsedatetime==2.6
pytz==2024.1
requests==2.31.0
groq==0.9.0
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
pytest==8.1.1
httpx==0.27.0
